feat(tools): install the CLI so it runs without a checkout
The CLI living only in the repository made ~/Library/Rime — the directory
that actually travels between machines — depend on a librime build tree, and
failed the spec's own goal: restoring a machine should not require cloning
and building librime first just to run restore.

install copies the package and the build_copilot binary into
private/bin and records the source commit plus a sha256 per file, so status
reports divergence on every run. That mitigation is the point: the directory
this replaces rotted precisely because nothing ever compared it to anything.

The installed entry point's shebang is rewritten to the absolute path of the
interpreter that ran install. A bare `#!/usr/bin/env python3` resolves
through pyenv by the caller's working directory rather than the script's
location, so the installed copy worked only from inside the one directory
holding a .python-version — which is not standalone. install is the only
place that knows which interpreter has the dependencies.

Also carries the whole-branch review's fixes. Two mattered on the restore
path: librime never writes sync_dir into installation.yaml (it only reads it,
defaulting in memory — verified against deployment_tasks.cc), so the README
was wrong and the first command it prescribed died on a traceback; and
missing-in-vault blocked restore permanently, though plan_backup silently
skips files the user does not have, making the round trip of a partial backup
a permanent failure signal. The documented test command also failed on a
stock interpreter, which the suite now no longer needs.

+++ tools/rime_copilot/dictdb.py
@@ -64,7 +64,7 @@ def merge(loaded: Sequence["tuple[Source, list[Entry]]"]) -> list[Entry]:
     every ordinary entry while its own frequency order survives inside the
     boost. Replacing outright was measured to invert real frequencies.
     """
-    merged: dict[tuple[str, str], int] = {}
+    merged: dict[tuple[str, str], float] = {}
 
     for source, entries in loaded:
         if source.top:
@@ -107,7 +107,7 @@ def write_pairs(entries: Sequence[Entry], out_path: Path, max_per_key: float) ->
     total = 0
     with open(out_path, "w", encoding="utf-8") as out:
         block_first: str | None = None
-        block: dict[tuple[str, str], int] = {}
+        block: dict[tuple[str, str], float] = {}
 
         def flush() -> int:
             written = 0
